+++ lib/src/controllers/th2_file_edit_selection_controller.dart
@@ -1092,6 +1092,9 @@ abstract class TH2FileEditSelectionControllerBase with Store {
           originalLineSegment: endControlPoint.lineSegment,
           type: endControlPoint.type,
         );
+    _th2FileEditController.optionEditController.setOptionElementsType(
+      MPOptionElementType.lineSegment,
+    );
     _th2FileEditController.stateController.updateStatusBarMessage();
   }